% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoProcData-package.R
\docType{package}
\name{ontoProcData-package}
\alias{ontoProcData-package}
\alias{ontoProcData}
\title{A data package for ontoProc}
\description{
This package manages rda files of multiple ontologies that are used in the ontoProc package. These ontologies were originally downloaded as owl or obo files and converted into Rda files . The files were downloaded at various times but most of them were downloaded on June 22 2021.
}
